pyproject.toml
[tool.poetry]
name = "arasaka corp tg ui"
version = "0.1.0"
description = "Telegram user interface for Arasaka Corporation"
authors = ["Anton Vice <a1996nton@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
Hydrogram = "^0.1.4"
py-tgcalls = "^1.1.2"
torch = {url = "https://download.pytorch.org/whl/cpu/torch-2.1.2-cp39-none-macosx_11_0_arm64.whl"}
python-dotenv = "^1.0.0"
tgcrypto = "^1.2.5"
faster-whisper = "^0.10.0"
uvloop = "^0.19.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
